made login page first thing, not you are not signed in page

+++ roomBooking/roomBooking/views.py
@@ -9,32 +9,35 @@ from django.db.models import Q
 def setupHomePage(request): # calls room view, user view, bookings view to display on page
     # roomData = Room.objects.all()
     # userData = User.objects.all()
-    roomBookingData = RoomBooking.objects.all()
+    if request.user.is_authenticated:
+        roomBookingData = RoomBooking.objects.all()
 
-    time = localtime(now())
+        time = localtime(now())
     
-    numBookings = roomBookingData.count()
-    numMeetingsNow = roomBookingData.filter(date=time.date(), start__lte=time.time(), end__gte=time.time()).count()
-    numRoomsAvailable = roomBookingData.count() - numMeetingsNow
+        numBookings = roomBookingData.count()
+        numMeetingsNow = roomBookingData.filter(date=time.date(), start__lte=time.time(), end__gte=time.time()).count()
+        numRoomsAvailable = roomBookingData.count() - numMeetingsNow
 
-    # if the booking date has passed or the booking is today but the end time has passed
-    passedBookings = roomBookingData.filter(Q(date__lt=time.date()) | Q(date=time.date(), end__lt=time.time()))
-    passedBookings.delete()
+        # if the booking date has passed or the booking is today but the end time has passed
+        passedBookings = roomBookingData.filter(Q(date__lt=time.date()) | Q(date=time.date(), end__lt=time.time()))
+        passedBookings.delete()
     
-    deletedBookings = DeletedBookingModel.objects.filter(userName=request.user.username)
+        deletedBookings = DeletedBookingModel.objects.filter(userName=request.user.username)
 
 
-    context = {
-        # 'rooms' : roomData,
-        # 'users' : userData,
-        'roomBooking' : roomBookingData,
-        'numBookings' : numBookings,
-        'time' : time,
-        'numMeetingsNow': numMeetingsNow,
-        'numRoomsAvailable': numRoomsAvailable,
-        'deletedBookings' : deletedBookings,
-    }
-    return render(request, 'home.html', context)
+        context = {
+            # 'rooms' : roomData,
+            # 'users' : userData,
+            'roomBooking' : roomBookingData,
+            'numBookings' : numBookings,
+            'time' : time,
+            'numMeetingsNow': numMeetingsNow,
+            'numRoomsAvailable': numRoomsAvailable,
+            'deletedBookings' : deletedBookings,
+        }
+        return render(request, 'home.html', context)
+    else:
+        return redirect("login")
 
 def clearDeletedBookingsUser(request):
     DeletedBookingModel.objects.filter(userName=request.user.username).delete()
